More adventures; Card events for discard, transfer, reveal
New adventures: Moonlight, Despair, Favorable Current, Sabotage.
Fixed AI-in-adventure for Man Overboard.
Players can now select their own discard.
Card Decks will now start empty, but know how to build a basic deck.
The game now has the concept of a discard pile.  Not sure if this should be its own pile or if it should be the main Action pile, but it's its own thing for now.